Moved most of nested sampling machinery to super class
